feat(chat): sync chats and messages with backend + suggested questions
