+++ src/components/ChatStart.vue
@@ -1,0 +1,12 @@
+<template>
+  <div class="chat-start container-fluid d-flex flex-column justify-content-center align-items-center">
+    <div class="row w-100">
+      <div class="col-12">
+        <div class="chat-start-content">
+          <h1 class="chat-start-title">Welcome to Chatbot</h1>
+          <p>Select a chat to continure</p>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
